requirements.txt
pyyaml
setuptools
wheel
requests